protos: Update build to work with latest protoc
This moves the openoffload build to be based on the latest protoc
compiler. This involved fixing all kinds of errors with the proobuf
files themselves. It also involved moving to building the protobuf files
using docker rather than requiring a natve build.

The sample code needs to be fixed as well, but before I do that, I
wanted to propose this to see if the maintainers of this project are ok
with this direction. If this is ok, I will next update the PR to make
the sample code compile with the latest here.

Signed-off-by: Kyle Mestery <mestery@mestery.com>