framework: Implement golang client/server
This adds a refreshed golang-based client and server set of programs
into the framework directory. We also build and push docker images for
these client and server container images. Appropriate CI is added to
GitHub Actions as well in the form of a new `framework-golang.yml` set
of tests.

In the future, we can add the equivalent for the C++ based framework
programs here.

Signed-off-by: Kyle Mestery <mestery@mestery.com>