golang: Return error when session table full
As we're adding sessions into the session table during a AddSession()
call, we call next_session_id() to get a new session ID. If the
session table is full, next_session_id() will return zero. We should
handle that case by appending a SessionResponseError with an error code
of _REJECTED_SESSION_TABLE_FULL in these instances.

Signed-off-by: Kyle Mestery <mestery@mestery.com>