xdp: sessionOffload XDP progress
This commit adds in a simple eBPF/XDP program which will handle the
backend of the sessionOffload API. For now, we simply define the eBPF
map used for communication from the golang server and the eBPF/XDP
program. The XDP program skeleton program is taken from a simple example
program I've written in the past [1]. It will be fleshed out to actually
operate on the sessions in the eBPF Map.

Related to #68

[1] https://github.com/mestery/xdp-skeleton

Signed-off-by: Kyle Mestery <mestery@mestery.com>